Corrijo las operaciones del TDA
